Made spawns better
The point of this change is to add code to setup spawnpoints for later, and make it so that the .1s delay doesn't let people find out who the imposter is.
This change blocks sending role updates for the crewmate or imposter role, and instead sends the updates manually once the roles are figured out to ensure that the crewmate/imposter cannot be determined.